mcp_business_agent/requirements.txt
# MCP Business Automation Agent Dependencies

# Core PydanticAI and model providers (fixed version conflicts)
pydantic-ai[openai,anthropic]==0.0.14
pydantic>=2.10
pydantic-settings>=2.6.1

# HTTP client for MCP communication
httpx==0.28.1
requests==2.32.3

# Environment variable management
python-dotenv==1.0.1

# Data validation and serialization
email-validator==2.2.0

# OAuth and security
cryptography==43.0.3

# Async support
asyncio-mqtt==0.16.2
aiofiles==24.1.0

# Testing framework
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Development tools
black==24.10.0
isort==5.13.2
mypy==1.13.0
ruff==0.8.4

# CLI interface
click==8.1.7
rich==13.9.4

# Logging and monitoring (optional)
structlog==24.4.0
loguru==0.7.2

# Production deployment (optional)
uvicorn==0.34.0
fastapi==0.115.6
gunicorn==23.0.0

# Additional utilities
python-multipart==0.0.12
jinja2==3.1.4
markupsafe==2.1.5